Restful tasks continued
Objectives:
Have an Angular app display the data it received from an API in the html
Continue building off your Restful Tasks assignment, which should currently have your service making all the get requests to your server. Now, have the root component invoke the methods in the service that make the requests to the server. Also, have the root component subscribe to the observables that are returned.

Next, have the root component store the data it receives in its own attributes.

Finally, you are set up to use the data you have in the root component to manipulate the DOM. Use structural directives to display the data in the HTML.