Restful Tasks Interactive
Objectives:
Have an Angular app make an http request to an API when the user clicks a button
Alter the DOM based on user interaction without refreshing the page
Right now, your Restful Tasks project is fetching data once the project is built. Let's give the power to the user! Alter your Restful Tasks assignment so that the user may click a button to fetch all the tasks and see them displayed in the browser. This happens without refreshing the page!

 Each task has a show button, which the user may click to view the details of the selected task below the main list. Selecting a different task will change the task being displayed. Again, this happens without refreshing the page.