Restful Tasks CRUD
Using your Task API assignment from the Mongo course, use Angular to complete the front-end portion of the API. We should be able to create a new task, view all tasks, update tasks, and delete tasks. Upon loading the page, have all the tasks displayed. When a new task is created, it should be appended to the list of tasks and the form should be cleared out. When a task is edited, have the edited task replace the task in the list. None of these actions should cause a refresh of the page.